backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "earthengine-api>=1.7.14",
    "fastapi>=0.129.0",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.34",
    "uvicorn>=0.41.0",
    "google-genai>=0.3.0",
    "datacommons_client>=2.0.0",

]
